changed datatypes to double

+++ sqllab/createtable.sql
@@ -1,9 +1,9 @@
 DROP TABLE IF EXISTS earthquakes;
 CREATE TABLE earthquakes (
-  latitude float,
-  longitude float,
-  depth float,
-  magnitude float,
+  latitude double,
+  longitude double,
+  depth double,
+  magnitude double,
   magnitudeType text,
   id text,
   locationSource text
